modify: 검색바필터 컴포넌트로 분리
